Add changable listening port and logging on server
The server port number is now changable so that client can be able to connect to the same computer on another port so that it can act as another channel. The messages that get sent to the server are now logged in a file so that they had be looked at later. This is important for security and user managment purposes

Server:
+ can change the port number on the server from the default
+ added logging